# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NIAT Campus Visit Manager API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export interface SuccessResponse {
  success: boolean;
  message: string;
}

export interface LoginRequest {
  phoneNumber: string;
  password: string;
}

export type UserInfoRole = (typeof UserInfoRole)[keyof typeof UserInfoRole];

export const UserInfoRole = {
  admin: "admin",
  sales: "sales",
} as const;

export interface UserInfo {
  id: number;
  phoneNumber: string;
  role: UserInfoRole;
}

export interface LoginResponse {
  token: string;
  user: UserInfo;
}

export interface SearchUserRequest {
  phoneNumber: string;
}

export type SearchUserResponseStudentInfo = {
  name?: string;
  email?: string;
  phone?: string;
  [key: string]: unknown;
};

export interface SearchUserResponse {
  isNewUser: boolean;
  userId?: string;
  applicationId?: string;
  studentInfo?: SearchUserResponseStudentInfo;
}

export interface GetCompletionRequest {
  applicationId: string;
  userId: string;
}

export interface GetCompletionResponse {
  bookedCampusVisit: number;
  visitedCampus: number;
}

export interface MarkVisitedRequest {
  userId: string;
  applicationId: string;
  phoneNumber: string;
}

export interface GenerateLinkRequest {
  userId: string;
  applicationId: string;
  phoneNumber: string;
}

export interface GenerateLinkResponse {
  redirectUrl: string;
  success: boolean;
}

export interface CreateSalesRequest {
  phoneNumber: string;
  password: string;
}

export interface SalesUsersResponse {
  users: UserInfo[];
}

export type AuditLogActionType =
  (typeof AuditLogActionType)[keyof typeof AuditLogActionType];

export const AuditLogActionType = {
  DIRECT_LINK_GENERATED: "DIRECT_LINK_GENERATED",
  CAMPUS_MARKED: "CAMPUS_MARKED",
} as const;

export interface AuditLog {
  id: number;
  actionType: AuditLogActionType;
  performedBy: string;
  performerPhone: string;
  targetUserId?: string;
  applicationId?: string;
  phoneNumber?: string;
  timestamp: string;
}

export interface LogsResponse {
  logs: AuditLog[];
  total: number;
  page: number;
  limit: number;
}

export type GetLogsParams = {
  phoneNumber?: string;
  performedBy?: string;
  page?: number;
  limit?: number;
};