# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NIAT Campus Visit Manager API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Login
 */
export const LoginBody = zod.object({
  phoneNumber: zod.string(),
  password: zod.string(),
});

export const LoginResponse = zod.object({
  token: zod.string(),
  user: zod.object({
    id: zod.number(),
    phoneNumber: zod.string(),
    role: zod.enum(["admin", "sales"]),
  }),
});

/**
 * @summary Search user by phone number
 */
export const SearchUserBody = zod.object({
  phoneNumber: zod.string(),
});

export const SearchUserResponse = zod.object({
  isNewUser: zod.boolean(),
  userId: zod.string().optional(),
  applicationId: zod.string().optional(),
  studentInfo: zod
    .object({
      name: zod.string().optional(),
      email: zod.string().optional(),
      phone: zod.string().optional(),
    })
    .optional(),
});

/**
 * @summary Get campus visit completion status
 */
export const GetCompletionBody = zod.object({
  applicationId: zod.string(),
  userId: zod.string(),
});

export const GetCompletionResponse = zod.object({
  bookedCampusVisit: zod.number(),
  visitedCampus: zod.number(),
});

/**
 * @summary Mark campus as visited (sales only)
 */
export const MarkVisitedBody = zod.object({
  userId: zod.string(),
  applicationId: zod.string(),
  phoneNumber: zod.string(),
});

export const MarkVisitedResponse = zod.object({
  success: zod.boolean(),
  message: zod.string(),
});

/**
 * @summary Generate direct visit link
 */
export const GenerateLinkBody = zod.object({
  userId: zod.string(),
  applicationId: zod.string(),
  phoneNumber: zod.string(),
});

export const GenerateLinkResponse = zod.object({
  redirectUrl: zod.string(),
  success: zod.boolean(),
});

/**
 * @summary Create a sales user (admin only)
 */
export const CreateSalesBody = zod.object({
  phoneNumber: zod.string(),
  password: zod.string(),
});

export const CreateSalesResponse = zod.object({
  success: zod.boolean(),
  message: zod.string(),
});

/**
 * @summary Get all sales users (admin only)
 */
export const GetSalesUsersResponse = zod.object({
  users: zod.array(
    zod.object({
      id: zod.number(),
      phoneNumber: zod.string(),
      role: zod.enum(["admin", "sales"]),
    }),
  ),
});

/**
 * @summary Get audit logs
 */
export const GetLogsQueryParams = zod.object({
  phoneNumber: zod.coerce.string().optional(),
  performedBy: zod.coerce.string().optional(),
  page: zod.coerce.number().optional(),
  limit: zod.coerce.number().optional(),
});

export const GetLogsResponse = zod.object({
  logs: zod.array(
    zod.object({
      id: zod.number(),
      actionType: zod.enum(["DIRECT_LINK_GENERATED", "CAMPUS_MARKED"]),
      performedBy: zod.string(),
      performerPhone: zod.string(),
      targetUserId: zod.string().optional(),
      applicationId: zod.string().optional(),
      phoneNumber: zod.string().optional(),
      timestamp: zod.string(),
    }),
  ),
  total: zod.number(),
  page: zod.number(),
  limit: zod.number(),
});