# lib/api-zod/src/generated/types/auditLog.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NIAT Campus Visit Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { AuditLogActionType } from "./auditLogActionType";

export interface AuditLog {
  id: number;
  actionType: AuditLogActionType;
  performedBy: string;
  performerPhone: string;
  targetUserId?: string;
  applicationId?: string;
  phoneNumber?: string;
  timestamp: string;
}