# lib/api-zod/src/generated/types/searchUserRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NIAT Campus Visit Manager API
 * OpenAPI spec version: 0.1.0
 */

export interface SearchUserRequest {
  phoneNumber: string;
}